simplyfing timetables
