+++ src/timetables/utc_timetables.rs
@@ -34,35 +34,35 @@
 // https://groups.google.com/d/forum/navitia
 // www.navitia.io
 
-use std::collections::{BTreeMap, HashMap};
-
-use super::{
-    generic_timetables::{self, Position, Timetable, Timetables, Trip, Vehicle, VehicleTimesError},
-    iters::{PositionsIter, TimetableIter, VehicleIter},
-    Stop, TimetablesIter,
-};
-
 use crate::{
-    loads_data::LoadsData,
+    loads_data::{Load, LoadsData},
     models::VehicleJourneyIdx,
-    time::{
-        days_patterns::{DaysInPatternIter, DaysPattern, DaysPatterns},
-        Calendar, DaysSinceDatasetStart, SecondsSinceDatasetUTCStart,
-        SecondsSinceTimezonedDayStart,
-    },
-    timetables::{FlowDirection, Timetables as TimetablesTrait, Types as TimetablesTypes},
+    request::generic_request::Mission,
+    time::days_patterns::{DaysInPatternIter, DaysPattern, DaysPatterns},
+    timetables::generic_timetables,
     RealTimeLevel,
 };
+
+use super::{
+    generic_timetables::{GenericTimetables, Trip, Vehicle, VehicleTimesError},
+    timetable_iters::{PositionsIter, TimetableIter},
+};
+use crate::{
+    time::{
+        Calendar, DaysSinceDatasetStart, SecondsSinceDatasetUTCStart,
+        SecondsSinceTimezonedDayStart, SecondsSinceUTCDayStart, TimezonesPatterns,
+    },
+    timetables::generic_timetables::{Position, Timetable},
+};
 use chrono::NaiveDate;
-use chrono_tz::Tz as TimeZone;
+use std::collections::{BTreeMap, HashMap};
 use tracing::log::error;
 
-use crate::loads_data::Load;
+use crate::timetables::{FlowDirection, Stop};
 
-pub type Time = SecondsSinceTimezonedDayStart;
-#[derive(Debug)]
-pub struct PeriodicTimetables {
-    timetables: Timetables<Time, Load, TimeZone, VehicleData>,
+pub struct UTCTimetables {
+    timetables: GenericTimetables<SecondsSinceUTCDayStart, Load, VehicleData>,
+    timezones_patterns: TimezonesPatterns,
 }
 
 #[derive(Debug, Clone)]
@@ -72,148 +72,125 @@ pub struct VehicleData {
     real_time_days_pattern: DaysPattern,
 }
 
-impl TimetablesTypes for PeriodicTimetables {
-    type Mission = Timetable;
-    type Position = Position;
-    type Trip = Trip;
-}
-
-impl TimetablesTrait for PeriodicTimetables {
-    fn new() -> Self {
+impl UTCTimetables {
+    pub fn new() -> Self {
         Self {
-            timetables: Timetables::new(),
+            timetables: GenericTimetables::new(),
+            timezones_patterns: TimezonesPatterns::new(),
         }
     }
 
-    fn nb_of_missions(&self) -> usize {
+    pub fn nb_of_missions(&self) -> usize {
         self.timetables.nb_of_timetables()
     }
 
-    fn mission_id(&self, mission: &Self::Mission) -> usize {
+    pub fn mission_id(&self, mission: &Mission) -> usize {
         mission.idx
     }
 
-    fn vehicle_journey_idx(&self, trip: &Self::Trip) -> VehicleJourneyIdx {
+    pub fn vehicle_journey_idx(&self, trip: &Trip) -> VehicleJourneyIdx {
         self.timetables
             .vehicle_data(&trip.vehicle)
             .vehicle_journey_idx
             .clone()
     }
 
-    fn stoptime_idx(&self, position: &Self::Position, _trip: &Self::Trip) -> usize {
+    pub fn stoptime_idx(&self, position: &Position, _trip: &Trip) -> usize {
         self.timetables.stoptime_idx(position)
     }
 
-    fn day_of(&self, trip: &Self::Trip) -> DaysSinceDatasetStart {
+    pub fn day_of(&self, trip: &Trip) -> DaysSinceDatasetStart {
         trip.day
     }
 
-    fn mission_of(&self, trip: &Self::Trip) -> Self::Mission {
+    pub fn mission_of(&self, trip: &Trip) -> Mission {
         self.timetables.timetable_of(&trip.vehicle)
     }
 
-    fn stop_at(&self, position: &Self::Position, mission: &Self::Mission) -> Stop {
+    pub fn stop_at(&self, position: &Position, mission: &Mission) -> super::Stop {
         *self.timetables.stop_at(position, mission)
     }
 
-    fn nb_of_trips(&self) -> usize {
+    pub fn nb_of_trips(&self) -> usize {
         self.timetables.nb_of_trips()
     }
 
-    fn is_upstream_in_mission(
+    pub fn is_upstream_in_mission(
         &self,
-        upstream: &Self::Position,
-        downstream: &Self::Position,
-        mission: &Self::Mission,
+        upstream: &Position,
+        downstream: &Position,
+        mission: &Mission,
     ) -> bool {
         self.timetables.is_upstream(upstream, downstream, mission)
     }
 
-    fn next_position(
-        &self,
-        position: &Self::Position,
-        mission: &Self::Mission,
-    ) -> Option<Self::Position> {
+    pub fn next_position(&self, position: &Position, mission: &Mission) -> Option<Position> {
         self.timetables.next_position(position, mission)
     }
 
-    fn previous_position(
-        &self,
-        position: &Self::Position,
-        mission: &Self::Mission,
-    ) -> Option<Self::Position> {
+    pub fn previous_position(&self, position: &Position, mission: &Mission) -> Option<Position> {
         self.timetables.previous_position(position, mission)
     }
 
-    fn arrival_time_of(
+    pub fn arrival_time_of(
         &self,
-        trip: &Self::Trip,
-        position: &Self::Position,
+        trip: &Trip,
+        position: &Position,
         calendar: &Calendar,
     ) -> (SecondsSinceDatasetUTCStart, Load) {
-        let (localtime, load) = self.timetables.arrival_time(&trip.vehicle, position);
-        let timetable = self.timetables.timetable_of(&trip.vehicle);
-        let timezone = self.timetables.timezone_data(&timetable);
-        let day = &trip.day;
-        let time = calendar.compose(day, localtime, timezone);
-        (time, *load)
+        let (time_in_day, load) = self.timetables.arrival_time(&trip.vehicle, position);
+        let time_utc = calendar.compose_utc(&trip.day, time_in_day);
+        (time_utc, *load)
     }
 
-    fn departure_time_of(
+    pub fn departure_time_of(
         &self,
-        trip: &Self::Trip,
-        position: &Self::Position,
+        trip: &Trip,
+        position: &Position,
         calendar: &Calendar,
     ) -> (SecondsSinceDatasetUTCStart, Load) {
-        let (localtime, load) = self.timetables.departure_time(&trip.vehicle, position);
-        let timetable = self.timetables.timetable_of(&trip.vehicle);
-        let timezone = self.timetables.timezone_data(&timetable);
-        let day = &trip.day;
-        let time = calendar.compose(day, localtime, timezone);
-        (time, *load)
+        let (time_in_day, load) = self.timetables.departure_time(&trip.vehicle, position);
+        let time_utc = calendar.compose_utc(&trip.day, time_in_day);
+        (time_utc, *load)
     }
 
-    fn debark_time_of(
+    pub fn debark_time_of(
         &self,
-        trip: &Self::Trip,
-        position: &Self::Position,
+        trip: &Trip,
+        position: &Position,
         calendar: &Calendar,
     ) -> Option<(SecondsSinceDatasetUTCStart, Load)> {
-        let has_localtime_load = self.timetables.debark_time(&trip.vehicle, position);
-        has_localtime_load.map(|(localtime, load)| {
-            let timetable = self.timetables.timetable_of(&trip.vehicle);
-            let timezone = self.timetables.timezone_data(&timetable);
+        let has_timeload_in_day = self.timetables.debark_time(&trip.vehicle, position);
+        has_timeload_in_day.map(|(time_in_day, load)| {
             let day = &trip.day;
-            let time = calendar.compose(day, localtime, timezone);
+            let time = calendar.compose_utc(day, time_in_day);
             (time, *load)
         })
     }
 
-    fn board_time_of(
+    pub fn board_time_of(
         &self,
-        trip: &Self::Trip,
-        position: &Self::Position,
+        trip: &Trip,
+        position: &Position,
         calendar: &Calendar,
     ) -> Option<(SecondsSinceDatasetUTCStart, Load)> {
-        let has_localtime_load = self.timetables.board_time(&trip.vehicle, position);
-        has_localtime_load.map(|(localtime, load)| {
-            let timetable = self.timetables.timetable_of(&trip.vehicle);
-            let timezone = self.timetables.timezone_data(&timetable);
+        let has_timeload_in_day = self.timetables.board_time(&trip.vehicle, position);
+        has_timeload_in_day.map(|(time_in_day, load)| {
             let day = &trip.day;
-            let time = calendar.compose(day, localtime, timezone);
+            let time = calendar.compose_utc(day, time_in_day);
             (time, *load)
         })
     }
 
-    fn earliest_trip_to_board_at(
+    pub fn earliest_trip_to_board_at(
         &self,
         waiting_time: &SecondsSinceDatasetUTCStart,
-        mission: &Self::Mission,
-        position: &Self::Position,
+        mission: &Mission,
+        position: &Position,
         real_time_level: &RealTimeLevel,
         calendar: &Calendar,
         days_patterns: &DaysPatterns,
-    ) -> Option<(Self::Trip, SecondsSinceDatasetUTCStart, Load)> {
+    ) -> Option<(Trip, SecondsSinceDatasetUTCStart, Load)> {
         self.earliest_filtered_trip_to_board_at(
             waiting_time,
             mission,
@@ -225,43 +202,28 @@ impl TimetablesTrait for PeriodicTimetables {
         )
     }
 
-    fn earliest_filtered_trip_to_board_at<Filter>(
+    pub fn earliest_filtered_trip_to_board_at<Filter>(
         &self,
         waiting_time: &SecondsSinceDatasetUTCStart,
-        mission: &Self::Mission,
-        position: &Self::Position,
+        mission: &Mission,
+        position: &Position,
         real_time_level: &RealTimeLevel,
         filter: Filter,
         calendar: &Calendar,
         days_patterns: &DaysPatterns,
-    ) -> Option<(Self::Trip, SecondsSinceDatasetUTCStart, Load)>
+    ) -> Option<(Trip, SecondsSinceDatasetUTCStart, Load)>
     where
         Filter: Fn(&VehicleJourneyIdx) -> bool,
     {
-        let has_earliest_and_latest_board_time =
-            self.timetables.earliest_and_latest_board_time(position);
+        let decompositions = calendar.decompositions_utc(waiting_time);
 
-        // if there is no earliest/latest board time, it means that this position cannot be boarded
-        // and we return None
-        let (_earliest_board_time_in_day, _latest_board_time_in_day) =
-            has_earliest_and_latest_board_time.map(|(earliest, latest)| (earliest, latest))?;
-
-        let timezone = self.timetables.timezone_data(mission);
-
-        let decompositions = calendar.decompositions(
-            waiting_time,
-            timezone,
-            SecondsSinceTimezonedDayStart::max(),
-            SecondsSinceTimezonedDayStart::min(),
-            // *latest_board_time_in_day,
-            // *earliest_board_time_in_day,
-        );
-        let mut best_vehicle_day_and_its_arrival_timeload_at_next_position: Option<(
+        let mut best_vehicle_day_and_its_arrival_time_at_next_position: Option<(
             Vehicle,
             DaysSinceDatasetStart,
             SecondsSinceDatasetUTCStart,
             Load,
         )> = None;
+
         for (waiting_day, waiting_time_in_day) in decompositions {
             let has_vehicle = self.timetables.earliest_filtered_vehicle_to_board(
                 &waiting_time_in_day,
@@ -278,25 +240,25 @@ impl TimetablesTrait for PeriodicTimetables {
             );
             if let Some((vehicle, arrival_time_in_day_at_next_stop, load)) = has_vehicle {
                 let arrival_time_at_next_stop =
-                    calendar.compose(&waiting_day, arrival_time_in_day_at_next_stop, timezone);
+                    calendar.compose_utc(&waiting_day, arrival_time_in_day_at_next_stop);
 
                 if let Some((_, _, best_arrival_time, best_load)) =
-                    &best_vehicle_day_and_its_arrival_timeload_at_next_position
+                    &best_vehicle_day_and_its_arrival_time_at_next_position
                 {
                     if arrival_time_at_next_stop < *best_arrival_time
                         || (arrival_time_at_next_stop == *best_arrival_time && load < best_load)
                     {
-                        best_vehicle_day_and_its_arrival_timeload_at_next_position =
+                        best_vehicle_day_and_its_arrival_time_at_next_position =
                             Some((vehicle, waiting_day, arrival_time_at_next_stop, *load));
                     }
                 } else {
-                    best_vehicle_day_and_its_arrival_timeload_at_next_position =
+                    best_vehicle_day_and_its_arrival_time_at_next_position =
                         Some((vehicle, waiting_day, arrival_time_at_next_stop, *load));
                 }
             }
         }
 
-        best_vehicle_day_and_its_arrival_timeload_at_next_position.map(
+        best_vehicle_day_and_its_arrival_time_at_next_position.map(
             |(vehicle, day, arrival_time_at_next_stop, load)| {
                 let trip = Trip { vehicle, day };
                 (trip, arrival_time_at_next_stop, load)
@@ -304,15 +266,15 @@ impl TimetablesTrait for PeriodicTimetables {
         )
     }
 
-    fn latest_trip_that_debark_at(
+    pub fn latest_trip_that_debark_at(
         &self,
         time: &SecondsSinceDatasetUTCStart,
-        mission: &Self::Mission,
-        position: &Self::Position,
+        mission: &Mission,
+        position: &Position,
         real_time_level: &RealTimeLevel,
         calendar: &Calendar,
         days_patterns: &DaysPatterns,
-    ) -> Option<(Self::Trip, SecondsSinceDatasetUTCStart, Load)> {
+    ) -> Option<(Trip, SecondsSinceDatasetUTCStart, Load)> {
         self.latest_filtered_trip_that_debark_at(
             time,
             mission,
@@ -324,37 +286,20 @@ impl TimetablesTrait for PeriodicTimetables {
         )
     }
 
-    fn latest_filtered_trip_that_debark_at<Filter>(
+    pub fn latest_filtered_trip_that_debark_at<Filter>(
         &self,
         time: &SecondsSinceDatasetUTCStart,
-        mission: &Self::Mission,
-        position: &Self::Position,
+        mission: &Mission,
+        position: &Position,
         real_time_level: &RealTimeLevel,
         filter: Filter,
         calendar: &Calendar,
         days_patterns: &DaysPatterns,
-    ) -> Option<(Self::Trip, SecondsSinceDatasetUTCStart, Load)>
+    ) -> Option<(Trip, SecondsSinceDatasetUTCStart, Load)>
     where
         Filter: Fn(&VehicleJourneyIdx) -> bool,
     {
-        let has_earliest_and_latest_debark_time =
-            self.timetables.earliest_and_latest_debark_time(position);
-
-        // if there is no earliest/latest debark time, it means that this position cannot be debarked
-        // and we return None
-        let (_earliest_debark_time_in_day, _latest_debark_time_in_day) =
-            has_earliest_and_latest_debark_time?;
-
-        let timezone = self.timetables.timezone_data(mission);
-
-        let decompositions = calendar.decompositions(
-            time,
-            timezone,
-            SecondsSinceTimezonedDayStart::max(),
-            SecondsSinceTimezonedDayStart::min(),
-            // *latest_debark_time_in_day,
-            // *earliest_debark_time_in_day,
-        );
+        let decompositions = calendar.decompositions_utc(time);
         let mut best_vehicle_day_and_its_departure_time_at_previous_position: Option<(
             Vehicle,
             DaysSinceDatasetStart,
@@ -376,15 +321,13 @@ impl TimetablesTrait for PeriodicTimetables {
                 },
             );
             if let Some((vehicle, departure_time_in_day_at_previous_stop, load)) = has_vehicle {
-                let departure_time_at_previous_stop = calendar.compose(
-                    &waiting_day,
-                    departure_time_in_day_at_previous_stop,
-                    timezone,
-                );
+                let departure_time_at_previous_stop =
+                    calendar.compose_utc(&waiting_day, departure_time_in_day_at_previous_stop);
+
                 if let Some((_, _, best_departure_time, best_load)) =
                     &best_vehicle_day_and_its_departure_time_at_previous_position
                 {
-                    if departure_time_at_previous_stop > *best_departure_time
+                    if departure_time_at_previous_stop >= *best_departure_time
                         || (departure_time_at_previous_stop == *best_departure_time
                             && load < best_load)
                     {
@@ -406,7 +349,7 @@ impl TimetablesTrait for PeriodicTimetables {
         )
     }
 
-    fn insert<'date, Stops, Flows, BoardTimes, DebarkTimes>(
+    pub fn insert<Stops, Flows, BoardTimes, DebarkTimes>(
         &mut self,
         stops: Stops,
         flows: Flows,
@@ -419,7 +362,7 @@ impl TimetablesTrait for PeriodicTimetables {
         timezone: &chrono_tz::Tz,
         vehicle_journey_idx: &VehicleJourneyIdx,
         real_time_level: &RealTimeLevel,
-    ) -> Result<HashMap<Self::Mission, DaysPattern>, (VehicleTimesError, Vec<NaiveDate>)>
+    ) -> Result<HashMap<Timetable, DaysPattern>, (VehicleTimesError, Vec<NaiveDate>)>
     where
         Stops: Iterator<Item = Stop> + ExactSizeIterator + Clone,
         Flows: Iterator<Item = FlowDirection> + ExactSizeIterator + Clone,
@@ -436,7 +379,7 @@ impl TimetablesTrait for PeriodicTimetables {
         };
         for date in days_patterns.make_dates(days, calendar) {
             let loads = loads_data
-                .loads(vehicle_journey_idx, &date)
+                .loads(&vehicle_journey_idx.clone(), &date)
                 .unwrap_or(default_loads.as_slice());
             load_patterns_dates
                 .entry(loads)
@@ -445,99 +388,129 @@ impl TimetablesTrait for PeriodicTimetables {
         }
 
         for (_loads, dates) in load_patterns_dates.iter() {
-            let days_pattern = days_patterns.get_from_dates(dates.iter(), calendar);
-            if days_patterns.is_empty_pattern(&days_pattern) {
-                continue;
-            }
+            let all_days_pattern = days_patterns.get_from_dates(dates.iter(), calendar);
 
-            let inspect_result = generic_timetables::inspect(
-                flows.clone(),
-                board_times.clone(),
-                debark_times.clone(),
-            );
-            if let Err(err) = inspect_result {
-                let dates = days_patterns.make_dates(&days_pattern, calendar);
-                return Err((err, dates));
+            for (offset, timezone_days_pattern) in
+                self.timezones_patterns
+                    .fetch_or_insert(timezone, days_patterns, calendar)
+            {
+                let days_pattern =
+                    days_patterns.get_intersection(all_days_pattern, *timezone_days_pattern);
+
+                if days_patterns.is_empty_pattern(&days_pattern) {
+                    continue;
+                }
+
+                let apply_offset = |time_in_timezoned_day: SecondsSinceTimezonedDayStart| -> SecondsSinceUTCDayStart {
+                    time_in_timezoned_day.to_utc(offset)
+                };
+
+                let board_times = board_times.clone().map(apply_offset);
+                let debark_times = debark_times.clone().map(apply_offset);
+
+                let inspect_result = generic_timetables::inspect(
+                    flows.clone(),
+                    board_times.clone(),
+                    debark_times.clone(),
+                );
+                if let Err(err) = inspect_result {
+                    let dates = days_patterns.make_dates(&days_pattern, calendar);
+                    return Err((err, dates));
+                }
             }
         }
 
         let mut result = HashMap::new();
 
         for (loads, dates) in load_patterns_dates.into_iter() {
-            let days_pattern = days_patterns.get_from_dates(dates.iter(), calendar);
-            if days_patterns.is_empty_pattern(&days_pattern) {
-                continue;
-            }
+            let all_days_pattern = days_patterns.get_from_dates(dates.iter(), calendar);
 
-            let (base_days_pattern, real_time_days_pattern) = match real_time_level {
-                RealTimeLevel::Base => (days_pattern, days_pattern),
-                RealTimeLevel::RealTime => (days_patterns.empty_pattern(), days_pattern),
-            };
+            for (offset, timezone_days_pattern) in
+                self.timezones_patterns
+                    .fetch_or_insert(timezone, days_patterns, calendar)
+            {
+                let days_pattern =
+                    days_patterns.get_intersection(all_days_pattern, *timezone_days_pattern);
 
-            let vehicle_data = VehicleData {
-                vehicle_journey_idx: vehicle_journey_idx.clone(),
-                base_days_pattern,
-                real_time_days_pattern,
-            };
-            let insert_result = self.timetables.insert(
-                stops.clone(),
-                flows.clone(),
-                board_times.clone(),
-                debark_times.clone(),
-                loads.iter().cloned(),
-                *timezone,
-                vehicle_data,
-            );
-            match insert_result {
-                Ok(mission) => {
-                    let pattern = result
-                        .entry(mission)
-                        .or_insert_with(|| days_patterns.empty_pattern());
-                    *pattern = days_patterns.get_union(*pattern, days_pattern);
+                if days_patterns.is_empty_pattern(&days_pattern) {
+                    continue;
                 }
-                Err(times_error) => {
-                    // this should not happen, since we inspect the times above
-                    // an returns early with an error if insertion should fail.
-                    // Let's log an error if this happens anyway
-                    error!(
-                        "An error occured while inserting a vehicle {:?}",
-                        times_error
-                    )
+
+                let (base_days_pattern, real_time_days_pattern) = match real_time_level {
+                    RealTimeLevel::Base => (days_pattern, days_pattern),
+                    RealTimeLevel::RealTime => (days_patterns.empty_pattern(), days_pattern),
+                };
+
+                let vehicle_data = VehicleData {
+                    vehicle_journey_idx: vehicle_journey_idx.clone(),
+                    base_days_pattern,
+                    real_time_days_pattern,
+                };
+
+                let apply_offset = |time_in_timezoned_day: SecondsSinceTimezonedDayStart| -> SecondsSinceUTCDayStart {
+                    time_in_timezoned_day.to_utc(offset)
+                };
+
+                let insert_result = self.timetables.insert(
+                    stops.clone(),
+                    flows.clone(),
+                    board_times.clone().map(apply_offset),
+                    debark_times.clone().map(apply_offset),
+                    loads.iter().cloned(),
+                    vehicle_data,
+                );
+                match insert_result {
+                    Ok(mission) => {
+                        let pattern = result
+                            .entry(mission)
+                            .or_insert_with(|| days_patterns.empty_pattern());
+                        *pattern = days_patterns.get_union(*pattern, days_pattern);
+                    }
+                    Err(times_error) => {
+                        // this should not happen, since we inspect the times above
+                        // an returns early with an error if insertion should fail.
+                        // Let's log an error if this happens anyway
+                        error!(
+                            "An error occured while inserting a vehicle. {:?}",
+                            times_error
+                        )
+                    }
                 }
             }
         }
-
         Ok(result)
     }
 
-    fn remove(
+    pub fn remove(
         &mut self,
-        mission: &Self::Mission,
+        timetable: &Timetable,
         day: &DaysSinceDatasetStart,
         vehicle_journey_idx: &VehicleJourneyIdx,
         real_time_level: &RealTimeLevel,
         _calendar: &Calendar,
         days_patterns: &mut DaysPatterns,
     ) {
-        let timetable_data = self.timetables.timetable_data_mut(mission);
+        let timetable_data = self.timetables.timetable_data_mut(timetable);
 
-        let nb_vehicle_updated = timetable_data.update_vehicles_data(|vehicle_data| {
-            let days_pattern = match real_time_level {
-                RealTimeLevel::Base => &mut vehicle_data.base_days_pattern,
-                RealTimeLevel::RealTime => &mut vehicle_data.real_time_days_pattern,
-            };
-            if vehicle_data.vehicle_journey_idx == *vehicle_journey_idx
-                && days_patterns.is_allowed(days_pattern, day)
-            {
-                *days_pattern = days_patterns
-                    .get_pattern_without_day(*days_pattern, day)
-                    .unwrap(); // unwrap is safe, because we check above that
-                               // vehicle_data.days_pattern contains day
-                true
-            } else {
-                false
-            }
-        });
+        let nb_vehicle_updated =
+            timetable_data.update_vehicles_data(|vehicle_data: &mut VehicleData| {
+                let days_pattern = match real_time_level {
+                    RealTimeLevel::Base => &mut vehicle_data.base_days_pattern,
+                    RealTimeLevel::RealTime => &mut vehicle_data.real_time_days_pattern,
+                };
+                if vehicle_data.vehicle_journey_idx == *vehicle_journey_idx
+                    && days_patterns.is_allowed(days_pattern, day)
+                {
+                    *days_pattern = days_patterns
+                        .get_pattern_without_day(*days_pattern, day)
+                        .unwrap(); // unwrap is safe, because we check above that
+                                   // days_patterns.is_allowed(&days_pattern, &day)
+                    true
+                } else {
+                    false
+                }
+            });
+
         if nb_vehicle_updated != 1 {
             error!("Updated {} vehicle during removal of one (vehicle_journey_idx, real_time_level, day).", nb_vehicle_updated);
         }
@@ -553,67 +526,63 @@ impl TimetablesTrait for PeriodicTimetables {
             error!("Removed {} vehicle during removal of one (vehicle_journey_idx, real_time_level, day).", nb_vehicle_removed);
         }
     }
-}
 
-impl<'a> TimetablesIter<'a> for PeriodicTimetables {
-    type Positions = PositionsIter;
-
-    fn positions(&'a self, mission: &Self::Mission) -> Self::Positions {
+    pub fn positions<'a>(&'a self, mission: &Mission) -> PositionsIter {
         self.timetables.positions(mission)
     }
 
-    type Trips = TripsIter<'a>;
-
-    fn trips_of(
+    pub fn trips_of<'a>(
         &'a self,
-        mission: &Self::Mission,
+        mission: &Mission,
         real_time_level: &RealTimeLevel,
         days_patterns: &'a DaysPatterns,
-    ) -> Self::Trips {
+    ) -> TripsIter<'a> {
         TripsIter::new(self, mission, real_time_level, days_patterns)
     }
 
-    type Missions = TimetableIter;
-
-    fn missions(&'a self) -> Self::Missions {
+    pub fn missions<'a>(&'a self) -> TimetableIter {
         self.timetables.timetables()
     }
 }
 
 pub struct TripsIter<'a> {
-    periodic: &'a PeriodicTimetables,
+    utc_timetables: &'a UTCTimetables,
     current_vehicle_days: Option<(Vehicle, DaysInPatternIter<'a>)>,
-    vehicles_iter: VehicleIter,
+    vehicles_iter: super::timetable_iters::VehicleIter,
     real_time_level: RealTimeLevel,
     days_patterns: &'a DaysPatterns,
 }
 
 impl<'a> TripsIter<'a> {
     fn new(
-        periodic: &'a PeriodicTimetables,
-        timetable: &Timetable,
+        utc_timetables: &'a UTCTimetables,
+        timetable: &super::generic_timetables::Timetable,
         real_time_level: &RealTimeLevel,
         days_patterns: &'a DaysPatterns,
     ) -> Self {
-        let mut vehicles_iter = periodic.timetables.vehicles(timetable);
+        let mut vehicles_iter = utc_timetables.timetables.vehicles(timetable);
         let has_current_vehicle = vehicles_iter.next();
         let current_vehicle_days = has_current_vehicle.map(|vehicle| {
             let days_pattern = match real_time_level {
-                RealTimeLevel::Base => periodic.timetables.vehicle_data(&vehicle).base_days_pattern,
+                RealTimeLevel::Base => {
+                    utc_timetables
+                        .timetables
+                        .vehicle_data(&vehicle)
+                        .base_days_pattern
+                }
                 RealTimeLevel::RealTime => {
-                    periodic
+                    utc_timetables
                         .timetables
                         .vehicle_data(&vehicle)
                         .real_time_days_pattern
                 }
             };
-
             let days_iter = days_patterns.days_in_pattern(&days_pattern);
             (vehicle, days_iter)
         });
 
         Self {
-            periodic,
+            utc_timetables,
             current_vehicle_days,
             vehicles_iter,
             real_time_level: real_time_level.clone(),
@@ -641,13 +610,13 @@ impl<'a> Iterator for TripsIter<'a> {
                         self.current_vehicle_days = has_current_vehicle.map(|vehicle| {
                             let days_pattern = match self.real_time_level {
                                 RealTimeLevel::Base => {
-                                    self.periodic
+                                    self.utc_timetables
                                         .timetables
                                         .vehicle_data(&vehicle)
                                         .base_days_pattern
                                 }
                                 RealTimeLevel::RealTime => {
-                                    self.periodic
+                                    self.utc_timetables
                                         .timetables
                                         .vehicle_data(&vehicle)
                                         .real_time_days_pattern
